+++ src/components/cars/constants.js
@@ -1,3 +1,23 @@
-const PARAMS = {};
+import car from './assets/car.png';
 
-export default PARAMS;
+const PARAMS = {
+  INITIAL_COORDINATES: [6000, 6450],
+  IMAGES: {
+    PLAYER_CAR: {
+      id: nanoid(),
+      img: car,
+      frameSize: { frameWidth: 256, frameHeight: 256 },
+    },
+  },
+};
+
+const controlKeys = {
+  up: 'W',
+  down: 'S',
+  left: 'A',
+  rigth: 'D',
+  stop: 'SPACE',
+  action: 'ENTER',
+};
+
+export { PARAMS, controlKeys };