+++ src/components/index.js
@@ -1,11 +1,32 @@
 import Map from './map';
 import Player from './player';
-import Car from './cars';
-import OrangeCar from './cars/orange-car';
-import BlueCar from './cars/blue-car';
-import Ways from './ways';
+import Pistol from './weapons/pistol';
+import MachineGun from './weapons/machine gun';
+import Chaingun from './weapons/chaingun';
+import Bullet from './weapons/bullet';
+import Car from './cars/standard';
+import PoliceCar from './cars/police';
+import TaxiCar from './cars/taxi';
+import RacingCar from './cars/racing car';
+import Citizens from './bots/citizens';
+import Rednecks from './bots/rednecks';
+import Interactions from './interactions';
 
-const features = [Car, Player, Map, OrangeCar, BlueCar, Ways];
+const features = [
+  Car,
+  Citizens,
+  Rednecks,
+  PoliceCar,
+  TaxiCar,
+  RacingCar,
+  Player,
+  Map,
+  Interactions,
+  Pistol,
+  Chaingun,
+  MachineGun,
+  Bullet,
+];
 const featuresId = [];
 
 const featuresMap = features.reduce((acc, FeatureCreater) => {